modify DummyEstimator for testing
